Revert "Rework vent critters (port from Delta-V)"
