+++ Content.Server/StationEvents/Events/VentCrittersRule.cs
@@ -1,8 +1,4 @@
-using Content.Server.Antag;
-using Content.Server.GameTicking.Rules.Components;
-using Content.Server.Pinpointer;
 using Content.Server.StationEvents.Components;
-using Content.Shared.EntityTable;
 using Content.Shared.GameTicking.Components;
 using Content.Shared.Station.Components;
 using Content.Shared.Storage;
@@ -13,13 +9,6 @@ using Robust.Shared.Player;
 
 namespace Content.Server.StationEvents.Events;
 
-/// <summary>
-/// DeltaV: Reworked vent critters to spawn a number of mobs at a single telegraphed location.
-/// This gives players time to run away and let sec do their job.
-/// </summary>
-/// <remarks>
-/// This entire file is rewritten, ignore upstream changes.
-/// </remarks>
 public sealed class VentCrittersRule : StationEventSystem<VentCrittersRuleComponent>
 {
     /*
@@ -28,87 +17,60 @@ public sealed class VentCrittersRule : StationEventSystem<VentCrittersRuleCompon
      */
 
     [Dependency] private readonly AnnouncerSystem _announcer = default!;
-    [Dependency] private readonly AntagSelectionSystem _antag = default!;
-    [Dependency] private readonly EntityTableSystem _entityTable = default!;
-    [Dependency] private readonly ISharedPlayerManager _player = default!;
-    [Dependency] private readonly NavMapSystem _navMap = default!;
-    [Dependency] private readonly SharedTransformSystem _transform = default!;
-
-    private List<EntityCoordinates> _locations = new();
 
     protected override void Added(EntityUid uid, VentCrittersRuleComponent component, GameRuleComponent gameRule, GameRuleAddedEvent args)
     {
         base.Added(uid, component, gameRule, args);
 
-        PickLocation(component);
-        if (component.Location is not { } coords)
-        {
-            ForceEndSelf(uid, gameRule);
+        if (!TryGetRandomStation(out var station))
             return;
-        }
-
-        var mapCoords = _transform.ToMapCoordinates(coords);
-        if (!_navMap.TryGetNearestBeacon(mapCoords, out var beacon, out _))
-            return;
-
-        var nearest = beacon?.Comp?.Text!;
 
         _announcer.SendAnnouncement(
             _announcer.GetAnnouncementId(args.RuleId),
             Filter.Broadcast(),
-            "station-event-vent-creatures-start-announcement-deltav",
+            "station-event-vent-creatures-announcement",
             null,
-            Color.Gold,
-            null, null,
-            ("location", nearest)
+            Color.Gold
         );
     }
 
-    protected override void Ended(EntityUid uid, VentCrittersRuleComponent comp, GameRuleComponent gameRule, GameRuleEndedEvent args)
+    protected override void Started(EntityUid uid, VentCrittersRuleComponent component, GameRuleComponent gameRule, GameRuleStartedEvent args)
     {
-        base.Ended(uid, comp, gameRule, args);
+        base.Started(uid, component, gameRule, args);
 
-        if (comp.Location is not { } coords)
-            return;
-
-        var players = _antag.GetTotalPlayerCount(_player.Sessions);
-        var min = Math.Max(comp.Min, comp.Min * players / comp.PlayerRatio);
-        var max = Math.Max(comp.Max, comp.Max * players / comp.PlayerRatio);
-        var count = Math.Max(RobustRandom.Next(min, max), 1);
-        Log.Info($"Spawning {count} critters for {ToPrettyString(uid):rule}");
-
-        for (int i = 0; i < count; i++)
-        {
-            foreach (var spawn in _entityTable.GetSpawns(comp.Table))
-            {
-                Spawn(spawn, coords);
-            }
-        }
-
-        if (comp.SpecialEntries.Count == 0)
-            return;
-
-        // guaranteed spawn
-        var specialEntry = RobustRandom.Pick(comp.SpecialEntries);
-        Spawn(specialEntry.PrototypeId, coords);
-    }
-
-    private void PickLocation(VentCrittersRuleComponent component)
-    {
         if (!TryGetRandomStation(out var station))
             return;
 
         var locations = EntityQueryEnumerator<VentCritterSpawnLocationComponent, TransformComponent>();
-        _locations.Clear();
-        while (locations.MoveNext(out var uid, out _, out var transform))
+        var validLocations = new List<EntityCoordinates>();
+        while (locations.MoveNext(out _, out _, out var transform))
         {
             if (CompOrNull<StationMemberComponent>(transform.GridUid)?.Station == station)
             {
-                _locations.Add(transform.Coordinates);
+                validLocations.Add(transform.Coordinates);
+                foreach (var spawn in EntitySpawnCollection.GetSpawns(component.Entries, RobustRandom))
+                {
+                    Spawn(spawn, transform.Coordinates);
+                }
             }
         }
 
-        if (_locations.Count > 0)
-            component.Location = RobustRandom.Pick(_locations);
+        if (component.SpecialEntries.Count == 0 || validLocations.Count == 0)
+        {
+            return;
+        }
+
+        // guaranteed spawn
+        var specialEntry = RobustRandom.Pick(component.SpecialEntries);
+        var specialSpawn = RobustRandom.Pick(validLocations);
+        Spawn(specialEntry.PrototypeId, specialSpawn);
+
+        foreach (var location in validLocations)
+        {
+            foreach (var spawn in EntitySpawnCollection.GetSpawns(component.SpecialEntries, RobustRandom))
+            {
+                Spawn(spawn, location);
+            }
+        }
     }
 }
