--- Content.Server/_Impstation/StationEvents/Components/SilentVentCrittersRuleComponent.cs
@@ -1,17 +1,0 @@
-using Content.Server.StationEvents.Events;
-using Content.Shared.Storage;
-
-namespace Content.Server.StationEvents.Components;
-
-[RegisterComponent, Access(typeof(SilentVentCrittersRule))]
-public sealed partial class SilentVentCrittersRuleComponent : Component
-{
-    [DataField("entries")]
-    public List<EntitySpawnEntry> Entries = new();
-
-    /// <summary>
-    /// At least one special entry is guaranteed to spawn
-    /// </summary>
-    [DataField("specialEntries")]
-    public List<EntitySpawnEntry> SpecialEntries = new();
-}
