--- Resources/Locale/en-US/_DV/station-events/events/vent-critters.ftl
@@ -1,1 +1,0 @@
-﻿station-event-vent-creatures-start-announcement-deltav = Attention. A large influx of unknown life forms has been detected in ventilation systems near {$location}. All personnel should vacate the area immediately.
