+++ Resources/Prototypes/_Impstation/GameRules/goblins.yml
@@ -7,7 +7,7 @@
     duration: 1
     minimumPlayers: 15
     maxOccurrences: 1 # six goblins is a lot
-  - type: SilentVentCrittersRule
+  - type: VentCrittersRule
     specialEntries:
     - id: GoblinStowawaysVentSpawner
       prob: 0 # this ensures that only one spawner will be created
